test: move inspector binary-frame test to its own file
inspect.test.ts has pre-existing failures in the parameterized
--inspect URL suite that are unrelated to this change. Keep the
binary-frame test standalone so the file passes cleanly.
